Extract UnusedPrivateProperty
